Use Context#deleteFile for local/private storage
mASAPP Report is reporting the following error:
```Unsafe file delete
Category: Code Quality
CVSS: 4.7
Description: The use of the method &#39;java.io/File/delete&#39; deletes the file or directory denoted by
its abstract pathname. If this pathname denotes a directory, then the directory must be empty
in order to be deleted.
Impact: This may lead to information disclosure if the deleted file was stored in an external SD
card. The original file could be recovered by a third-party in an undesired manner.
Recommendations: Please make sure not to use &#39;file.delete&#39; to delete essential files. Use
&#39;Context.deleteFile&#39; for private files instead.
```